% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-workflows.R
\name{add_workflows}
\alias{add_workflows}
\title{Create a miter table by combining data with a list of workflows}
\usage{
add_workflows(x, workflows)
}
\arguments{
\item{x}{A data frame or a miter table containing all relevant variables.}

\item{workflows}{A \code{workflow} object or a list containing \code{workflow} objects.}
}
\value{
A nested tibble with extra class \code{miter_tbl} and the following added
columns:
\itemize{
\item \code{models} contains character strings that identify the models provided in the \code{workflows} argument.
\item \code{workflows} contains the workflows objects provided in the argument \code{workflows}
}
}
\description{
Create a miter table by combining data with a list of workflows
}
