% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdout-splits.R
\name{holdout_split}
\alias{holdout_split}
\title{Create holdout splits for each group}
\usage{
holdout_split(
  x,
  column = data,
  colname = "splits",
  prop = 3/4,
  strata = NULL,
  breaks = 4,
  pool = 0.1,
  ...
)
}
\arguments{
\item{x}{A data frame or miter table containing all relevant variables}

\item{prop}{Description}
}
\value{
A miter table with the following added columns:
\itemize{
\item \code{splits} a list of rsplit objects
}
}
\description{
Create holdout splits for each group
}
