% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{:=}
\alias{fit}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{generics}{\code{\link[generics]{fit}}}

  \item{rlang}{\code{\link[rlang:dyn-dots]{:=}}}
}}

